requirements.txt
streamlit
openai
requests
python-docx
docx2pdf
Pillow
diffusers
transformers
accelerate